+++ fdr-stack/src/main/java/com/dwarfeng/fdr/stack/handler/KeepHandler.java
@@ -34,8 +34,8 @@ public interface KeepHandler<D extends Data> extends Handler {
      * 更新数据。
      *
      * <p>
-     * 调用此方法时，需要保证 {@link Data#getHappenedDate()} 的值大于等于保持器中对应数据的发生时间。<br>
-     * 该规约通常由数据源保证。
+     * 调用此方法时，{@link Data#getHappenedDate()} 的值有可能小于保持器中对应数据的发生时间。<br>
+     * 不同的保持器对于该情形的处理方式不同，具体的处理方式由保持器的实现者决定。
      *
      * @param data 数据。
      * @throws HandlerException 处理器异常。
@@ -50,9 +50,9 @@ public interface KeepHandler<D extends Data> extends Handler {
      * 该规约通常由记录处理器保证。
      *
      * <p>
-     * 调用此方法时，需要保证参数列表中所有的数据的 {@link Data#getHappenedDate()}
-     * 的值大于等于保持器中对应数据的发生时间。<br>
-     * 该规约通常由数据源保证。
+     * 调用此方法时，需要保证参数列表中有可能存在数据的 {@link Data#getHappenedDate()} 的值小于保持器中对应数据的发生时间。
+     * <br>
+     * 不同的保持器对于该情形的处理方式不同，具体的处理方式由保持器的实现者决定。
      *
      * @param datas 数据组成的列表。
      * @throws HandlerException 处理器异常。